# Optimize the no-escape case with strpbrk