@k0kubun 2022-11-05
I actually reverted this idea on the other side myself https://github.com/ruby/ruby/commit/9af344a42129ea193d6dba92d62458c5f3429507. Closing for now.